requirements.txt
click==6.6
decorator==4.0.10
docker-py==1.9.0
Flask==0.11.1
flask-codemirror==1.0
Flask-WTF==0.13.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
websocket-client==0.37.0
Werkzeug==0.11.10
WTForms==2.1
